environments/gem_wordle/pyproject.toml
[project]
name = "gem_wordle"
description = "Gym-compatible Wordle environment powered by AxonRL's GEM framework"
tags = ["gem", "multi-turn", "reasoning", "game", "gym", "train", "eval"]
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "verifiers>=0.1.8",
    "gem-llm",
    "gymnasium",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["gem_wordle.py"]

[tool.verifiers.eval]
num_examples = 20
rollouts_per_example = 1